+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- A basic example bot implementation has been added in the `examples` directory
+  of the repository.
 
 ## [0.15.2] - 2023-06-23
 ### Added